Moved server to own python package
